Remove CPU resource limit from processor
I realized that setting a limit implies a corresponding request if
it's not already defined, which was unintentional. Instead of adding
an explicit request value, for now I'm just going to remove the config
altogether since these processors should be I/O bound for the most
part anyways. We can revisit in the future as needed.