Update raise-data image with workflow dependencies
This change adds a new requirements.txt that will be used to pin
dependencies for the raise-data Docker image which are separate
from those associated with the Python package in this repo. It also
updates the build to push a latest tag so we can rely on it for
workflows, etc.